fix: fixed the delay between the title an background load
By combining the two promises that were loading the title and background
seperatly and making the promise load both and then when resolved update
the description wich is almost instantaneous, and restarting the
animation it will have a smother loading. As mention before changing the
format of the image into a webp made the images lighter and faster to
load.